+++ src/OpenTelemetry.Exporter.Jaeger/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed `JaegerExporter` class and constructor from internal to public.
+  ([#1612](https://github.com/open-telemetry/opentelemetry-dotnet/issues/1612))
+
 * In `JaegerExporterOptions`: Exporter options now include a switch for
   Batch vs Simple exporter, and settings for batch exporting properties.